ajout d'un bonus de changement de taille de la barre

+++ CasseBrique/CasseBrique/Views/ShapeView.cs
@@ -30,10 +30,8 @@ namespace Breakout.Views
 
         public void Draw(SpriteBatch spriteBatch, GameTime gameTime)
         {
-            float scale = Shape.Size.Width / texture.Width;
-
+            float scale = (float)Shape.Size.Width / texture.Width;
             spriteBatch.Draw(this.Texture, Shape.Position, null, Color.White, 0, Vector2.Zero, scale, SpriteEffects.None, 1);   
-            //spriteBatch.Draw(this.Texture, Shape.Position, Color.White);
         }
     }
 }
